adding new acq function to trieste
