+++ docs/notebooks/multiobjective_constrained.py
@@ -176,40 +176,7 @@ model = {OBJECTIVE: objective_model, CONSTRAINT: constraint_model}
 # Here we utilize the [EHVI](https://link.springer.com/article/10.1007/s10898-019-00798-7): `ExpectedHypervolumeImprovement` acquisition function:
 
 # %%
-from trieste.acquisition.function import ExpectedConstrainedImprovement, expected_hv_improvement
-
-class ExpectedConstrainedHypervolumeImprovement(ExpectedConstrainedImprovement):
-
-    def __repr__(self) -> str:
-        """"""
-        return (
-            f"ExpectedConstrainedHypervolumeImprovement({self._objective_tag!r}, {self._constraint_builder!r},"
-            f" {self._min_feasibility_probability!r})"
-        )
-
-    def prepare_acquisition_function(self, datasets, models):
-        objective_model = models[self._objective_tag]
-        objective_dataset = datasets[self._objective_tag]
-
-        if len(objective_dataset) == 0:
-            raise ValueError(
-                "Expected hypervolume improvement is defined with respect to existing points in the objective"
-                " data, but the objective data is empty."
-            )
-
-        constraint_fn = self._constraint_builder.prepare_acquisition_function(datasets, models)
-        pof = constraint_fn(objective_dataset.query_points[:, None, ...])
-        is_feasible = pof >= self._min_feasibility_probability
-
-        if not tf.reduce_any(is_feasible):
-            return constraint_fn
-
-        objective_mean, _ = objective_model.predict(objective_dataset.query_points)
-
-        _pf = Pareto(tf.boolean_mask(objective_mean, is_feasible[:,0]))
-        _reference_pt = get_reference_point(_pf.front)
-        ehvi = expected_hv_improvement(objective_model, _pf, _reference_pt)
-        return lambda at: ehvi(at) * constraint_fn(at)
+from trieste.acquisition.function import ExpectedConstrainedHypervolumeImprovement
 
 pof = trieste.acquisition.ProbabilityOfFeasibility(threshold=Sim.threshold)
 echvi = ExpectedConstrainedHypervolumeImprovement(OBJECTIVE, pof.using(CONSTRAINT))
